Refactored FR-18: moved override to routes and schemas
