+++ Backend/schemas/override.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class OverrideRequest(BaseModel):
+    username: str
+    password: str
+    action: str
+
